Pretty much seperated stuff in components folder in 3 categories
