+++ src/components/common/buttons/addToCompare.js
