+++ src/components/common/buttons/compare.js
