+++ src/components/common/gauge/index.js
