+++ src/components/common/image-gallery/styles.jsx
