+++ src/components/sections/results/compare-results/index.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useParams } from 'react-router';
 import { Link } from 'react-router-dom';
 import { List, Card } from "antd";
-import { fetchSelectedCarDataByYear } from '../../../hooks/dataFetching';
+import { fetchSelectedCarDataByYear } from '../../../../hooks/dataFetching';
 import styled from 'styled-components';
 
 import ResultsStyled from '../home-results/styles';
