+++ src/components/sections/results/home-results/styles.jsx
@@ -2,7 +2,6 @@
 import styled from 'styled-components';
 
 const ResultsStyled = styled.div`
-
   // .home-results-title {
   //   margin-top: 10;
   //   margin-bottom: 10;
