+++ src/components/sections/search/compare-search/index.js
@@ -1,17 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import { fetchMakeData, fetchModelData, fetchYearData } from '../../../hooks/dataFetching';
+import { fetchMakeData, fetchModelData, fetchYearData } from '../../../../hooks/dataFetching';
 import { handleMakeChanges, 
         handleModelChanges, 
         handleYearChanges, 
         handleClear, 
         disableOtherDropdown, 
         disableYearDropdown, 
-      } from '../../../hooks/dropdownFunctions';
-import Dropdown from '../../dropdown';
-import HomeResults from '../../results/home-results';
+      } from '../../../../hooks/dropdownFunctions';
+import Dropdown from '../../../common/dropdown';
 import CompareResults from '../../results/compare-results';
 
-export default function Search({ searchTitle, id })  {
+export default function CompareSearch({ searchTitle })  {
   const [carMakes, setCarMakes] = useState([]);
   const [makeSelected, setMakeSelected] = useState("");
 
@@ -71,10 +70,8 @@ export default function Search({ searchTitle, id })  {
           data={carYears}/>
       </div>
       { modelSelected ? (
-        <HomeResults make = {makeSelected} model = {modelSelected} year = {yearSelected} />)
-      : id ? <CompareResults id={id} make = {makeSelected} model = {modelSelected} year = {yearSelected} />
-      : <></>
-      }
+        <CompareResults make = {makeSelected} model = {modelSelected} year = {yearSelected} />)
+      : <></>}
     </>
   );
 };
