+++ src/components/sections/search/home-search/index.js
@@ -1,16 +1,17 @@
 import React, { useState, useEffect } from 'react';
-import { fetchMakeData, fetchModelData, fetchYearData } from '../../../hooks/dataFetching';
+import { fetchMakeData, fetchModelData, fetchYearData } from '../../../../hooks/dataFetching';
 import { handleMakeChanges, 
         handleModelChanges, 
         handleYearChanges, 
         handleClear, 
         disableOtherDropdown, 
         disableYearDropdown, 
-      } from '../../../hooks/dropdownFunctions';
-import Dropdown from '../../dropdown';
+      } from '../../../../hooks/dropdownFunctions';
+import Dropdown from '../../../common/dropdown';
+import HomeResults from '../../results/home-results';
 import CompareResults from '../../results/compare-results';
 
-export default function CompareSearch({ searchTitle })  {
+export default function Search({ searchTitle, id })  {
   const [carMakes, setCarMakes] = useState([]);
   const [makeSelected, setMakeSelected] = useState("");
 
@@ -70,8 +71,10 @@ export default function CompareSearch({ searchTitle })  {
           data={carYears}/>
       </div>
       { modelSelected ? (
-        <CompareResults make = {makeSelected} model = {modelSelected} year = {yearSelected} />)
-      : <></>}
+        <HomeResults make = {makeSelected} model = {modelSelected} year = {yearSelected} />)
+      : id ? <CompareResults id={id} make = {makeSelected} model = {modelSelected} year = {yearSelected} />
+      : <></>
+      }
     </>
   );
 };
